requirements.txt
# Core dependencies
pyzabbix==1.3.0
pynetbox==7.3.3
redis==5.0.1
python-dotenv==1.0.0

# Telegram notifications
python-telegram-bot==20.7
requests==2.31.0

# Utils
urllib3==2.1.0